refactor: move devDependencies that end-users need installed to dependencies
This mostly just  moves a handful of devDependencies to dependencies. It
doesn't change any of the dependencies listed, or alter the versions of
those dependencies, but it does install one new one
(eslint-plugin-react-hooks) because I was getting warnings that it was
an ininstalled peer dependency of "eslint-config-airbnb".

A downstream user upgrading to this version of eslint-config-sparkbox could
end up with a different version of a package, if the version that they had
previously installed as a peer-dependency differs from the version that is
now declared as a dependency in this repo.

That being said, I don't *think* there's a risk of breaking changes. For
an example, if a downstream user was pulling in a newer version of
'eslint-plugin-react' which deprecated the 'react/require-extension'
rule, then eslint would have failed for them already. If it currently
works, then falling back to an earlier version of the package that still
has the same rule should work fine. I think.

Closes #25