chore: update dependencies
This follows the instructions in the airbinb/javascript repo
(https://github.com/airbnb/javascript/tree/master/packages/eslint-config-airbnb#eslint-config-airbnb-1)
to update this package to ESLint 7.X.X and ensure full compatibility of
all the companion plugins that depend on it.

Note: this did create some test failures that needed to be updated.
Consequently, this will likely result in a major version update.