Revert "refactor: move devDependencies that end-users need installed to dependencies"
This reverts commit ab37b7f2546cf524761dfaa2f3036b44cadc3d42.

Reverting this because it turns out that we can't actually declare
eslint plugins as dependencies. There's a feature request for this
that has been open since 2015: https://github.com/eslint/eslint/issues/3458

This is also described in the docs but I missed it because I mixed
up "sharable configs" with "plugins". More details on that here:
https://eslint.org/docs/developer-guide/shareable-configs#publishing-a-shareable-config

It looks like there is a "simple config" initiative in progress which
should improve this (https://github.com/eslint/eslint/issues/13481).
But until that's available, it's probably best to revert this to the
existing approach.